Introduce CodafBaseDetalheDTO and refactor services
Add a new CodafBaseDetalheDTO (with RetificacaoDTO and AnexoCodafDetalheDTO) to centralize shared Codaf detalhe fields. Update codaf-lista-presenca-service to import and extend the base DTO (removing its duplicated detail types) and adjust imports. Update codaf-suplementar-service to extend CodafBaseDetalheDTO and consolidate common fields into the new DTO. This reduces duplication and standardizes detail DTOs across services.
